pyproject.toml
[tool.poetry]
name = "hhparser"
version = "0.1.0"
description = ""
authors = ["Kostin Igor <grizzlygit777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
loguru = "^0.7.2"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.3.2"
pylint = "^3.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
